fix(audit): pass 8 — audio size vs bound texture, restart-scope gaps, file-size seams
The audio spectrum count published in the UBO was derived from the pending bar
count rather than the texture actually bound. Two reachable paths left a frame
advertising more bars than the texture was wide: a failed create() keeps the
previous smaller texture, and the dummy-channel early return skips the audio
block outright. audio.glsl gates barIndex on that count and then texelFetches,
which is undefined out of range. Derive it from the bound texture and re-arm
the uniforms after a successful resize, so it trails by a frame instead of
over-reporting.

Both shader registries moved from the constructor into init(): stop() resets
them, so a stop/init/start cycle left them null for the rest of the process
with the overlay service's borrows dead and every warm bake skipped.

The eight autotile shortcut handlers now track their connections, so a restart
no longer fires each one twice. They need their own list because start() calls
initializeAutotile() before connectLayoutSignals() clears the existing one.

uSurfaceOpacity is documented as always 1.0, but the daemon pushes the item's
live opacity so a host can fade the decoration. Correct the contract. The user
texture clamp claimed QRhi scale-fits an oversized image, which it does not, so
scale the source before upload.

Route ZoneShaderItem's include-path walk through the shared helper it was
duplicating, and hoist that helper into the header beside resolveZoneVertexPath.

daemon.h was over the 1150 ceiling; trim its longest doc essays. signals.cpp
went over too, so extract initializeAutotile() into daemon/autotile_init.cpp
along the seam the remediation plan already named for it.
