+++ docs/file-size-remediation.md
@@ -53,8 +53,8 @@ Grace-band headers that look like exceptions but are merely TOLERATED (single-cl
 headers under 1150, no action): `src/daemon/daemon.h` (1149), `src/daemon/overlayservice.h`
 (1137), `src/settings/controller/settingscontroller.h` (1149),
 `PhosphorZones/LayoutRegistry.h` (1119), `PhosphorSnapEngine/SnapEngine.h` (1079).
-Warning for daemon.h: the daemon.cpp split adds ~10 private declarations; trim its
-longest doc essays in the same PR to stay under 1150.
+Warning for daemon.h: it now sits exactly at 1150 after one round of doc-essay
+trimming, so any further declaration must be paid for by trimming again.
 
 `src/ui/PassiveOverlayShell.qml` (1047) is TOLERATED but flagged **near-exception**:
 C++ (`OverlayService`) writes properties directly onto its slot Items and inner
@@ -339,7 +339,7 @@ list; missing QML files in `qt6_add_qml_module` fail at runtime, not build time.
 | 1189 | libs/phosphor-tiles/src/pluau/pluau.luau | EXCEPTION |
 | 1160 | src/editor/EditorController.h | SPLIT (§2.6) |
 | 1149 | src/settings/controller/settingscontroller.h | TOLERATED |
-| 1149 | src/daemon/daemon.h | TOLERATED (watch: daemon.cpp split adds decls) |
+| 1150 | src/daemon/daemon.h | TOLERATED at the ceiling; doc essays trimmed once already, next addition must trim again |
 | 1137 | src/daemon/overlayservice.h | TOLERATED |
 | 1131 | src/config/settingsschema.cpp | TOLERATED (seam: per-domain append functions) |
 | 1130 | libs/phosphor-zones/src/layoutregistry_assignments.cpp | TOLERATED (easy seam: contextresolve vs CRUD) |
@@ -353,7 +353,7 @@ list; missing QML files in `qt6_add_qml_module` fail at runtime, not build time.
 | 1099 | libs/phosphor-config/src/jsonbackend.cpp | TOLERATED (trivial seam: JsonGroup) |
 | 1090 | src/daemon/daemon/start.cpp | TOLERATED (easy seam: shortcuts.cpp / virtual_screens.cpp) |
 | 1089 | tests/unit/settings/stores/test_profilestore.cpp | TOLERATED |
-| 1088 | src/daemon/daemon/signals.cpp | TOLERATED (easy seam: autotile_init.cpp) |
+| 625 | src/daemon/daemon/signals.cpp | DONE — initializeAutotile() extracted to daemon/autotile_init.cpp (568) |
 | 1087 | src/editor/qml/PropertyPanel.qml | TOLERATED (twinFormLayouts obstacle) |
 | 1080 | src/settings/stores/profilestore.cpp | TOLERATED (seam: CRUD partial) |
 | 1079 | libs/phosphor-snap-engine/include/PhosphorSnapEngine/SnapEngine.h | TOLERATED |
