+++ src/daemon/daemon.h
@@ -678,27 +678,20 @@ private:
     // stays valid across back-to-back resolves.
     PhosphorRules::RuleSet m_excludeRuleSet;
     std::unique_ptr<PhosphorZones::LayoutRegistry> m_layoutManager;
-    // Daemon-owned tile-algorithm registry. Replaces the old
-    // AlgorithmRegistry::instance() singleton — per-process ownership is
-    // the only shape that works once PlasmaZones becomes a plugin-based
-    // compositor/WM/shell (plugins can't share process-global state
-    // safely).
-    //
-    // ─── DECLARATION ORDER INVARIANT ─────────────────────────────────
-    // Every FactoryContext service the bundle borrows (m_layoutManager
-    // → IZoneLayoutRegistry, m_algorithmRegistry → ITileAlgorithmRegistry)
-    // MUST be declared before m_layoutSources so reverse-order member
-    // destruction tears the bundle (and its ZonesLayoutSource /
-    // AutotileLayoutSource children) down BEFORE the registries those
-    // children borrow. The LayoutSourceBundle contract
-    // (libs/phosphor-layout-api/.../LayoutSourceBundle.h) is explicit
-    // about this — violating the order produces dangling pointers in
-    // every source's destructor, hidden today only by Qt's signal
-    // auto-disconnect. Do not reorder these three lines without revisiting
-    // every source's destructor.
-    //
-    // Also declared before ScriptedAlgorithmLoader + AutotileEngine
-    // because both take a borrowed pointer to it in their constructor.
+    // Daemon-owned tile-algorithm registry, replacing the old
+    // AlgorithmRegistry::instance() singleton: plugins can't share
+    // process-global state safely, so the composition root owns it.
+    // DECLARATION ORDER INVARIANT: every FactoryContext service the bundle
+    // borrows (m_layoutManager → IZoneLayoutRegistry, m_algorithmRegistry
+    // → ITileAlgorithmRegistry) MUST precede m_layoutSources, so
+    // reverse-order destruction tears the bundle and its ZonesLayoutSource
+    // / AutotileLayoutSource children down before the registries they
+    // borrow. See the LayoutSourceBundle contract
+    // (libs/phosphor-layout-api/.../LayoutSourceBundle.h). Violating it
+    // dangles every source's destructor, hidden today only by Qt's signal
+    // auto-disconnect, so don't reorder these three lines without
+    // revisiting them. Must also precede ScriptedAlgorithmLoader and
+    // AutotileEngine, which borrow it in their constructors.
     std::unique_ptr<PhosphorTiles::AlgorithmRegistry> m_algorithmRegistry;
     // Manual layouts + autotile algorithms composed behind layoutSource().
     // The bundle owns all three objects so destruction is deterministic
@@ -730,33 +723,25 @@ private:
     /// adding a second path needs no new plumbing. Borrows nothing, so it sits
     /// outside the declaration-order block below.
     QHash<QString, PhosphorAnimation::Profile> m_rawJsonProfiles;
-    /// Per-daemon curve registry. Replaces the prior per-process
-    /// `CurveRegistry::instance()` singleton — composition roots own
-    /// their own.
-    ///
-    /// DECLARATION ORDER INVARIANT: must be declared BEFORE `m_settings`
-    /// (which takes a borrowed pointer to it in its constructor) and
-    /// BEFORE `m_curveLoader` / `m_profileLoader` (which also borrow).
-    /// Reverse-order destruction then tears every consumer down before
-    /// the registry itself, guaranteeing no UAF on the Settings /
-    /// loader teardown paths. Also reset from `PhosphorCurve::s_registry`
-    /// in `~Daemon` before teardown so the QML static helper doesn't
-    /// dangle into freed memory on process shutdown or successive
-    /// Daemon constructions in tests.
+    /// Per-daemon curve registry, replacing the `CurveRegistry::instance()`
+    /// singleton so each composition root owns its own.
+    /// DECLARATION ORDER INVARIANT: must precede `m_settings`,
+    /// `m_curveLoader` and `m_profileLoader`, all of which borrow it, so
+    /// reverse-order destruction tears every consumer down first and no
+    /// Settings / loader teardown path can UAF. Also cleared from
+    /// `PhosphorCurve::s_registry` in `~Daemon`, so the QML static helper
+    /// can't dangle on shutdown or across successive Daemon
+    /// constructions in tests.
     PhosphorAnimation::CurveRegistry m_curveRegistry;
-    /// Per-daemon profile registry. Replaces the prior process-global
-    /// `PhosphorProfileRegistry::instance()` singleton — composition
-    /// roots own their own and publish via `setDefaultRegistry` so QML
-    /// callsites resolve through the same instance the daemon
-    /// populates from Settings + ProfileLoader.
-    ///
-    /// DECLARATION ORDER INVARIANT: must be declared BEFORE
-    /// `m_overlayService` (which takes a reference into its
-    /// SurfaceAnimator) and BEFORE `m_profileLoader` (which holds a
-    /// reference). Reverse-order destruction then tears the consumers
-    /// down before the registry itself, guaranteeing no UAF on the
-    /// service / loader teardown paths. The `setDefaultRegistry(nullptr)`
-    /// call in `stop()` clears the QML static handle before teardown.
+    /// Per-daemon profile registry, replacing the
+    /// `PhosphorProfileRegistry::instance()` singleton. Published via
+    /// `setDefaultRegistry` so QML callsites resolve through the same
+    /// instance the daemon populates from Settings + ProfileLoader.
+    /// DECLARATION ORDER INVARIANT: must precede `m_overlayService` (which
+    /// references it from its SurfaceAnimator) and `m_profileLoader`, so
+    /// reverse-order destruction tears the consumers down first and no
+    /// service / loader teardown path can UAF. `stop()` calls
+    /// `setDefaultRegistry(nullptr)` to clear the QML static handle.
     PhosphorAnimation::PhosphorProfileRegistry m_profileRegistry;
     /// Per-daemon QtQuickClock manager — replaces the prior process-
     /// global `QtQuickClockManager::instance()` singleton. Published via
@@ -803,25 +788,24 @@ private:
     /// Wayland client, so it watches and pushes the resolved boolean to the effect
     /// over D-Bus (SettingsAdaptor::sessionIdleChanged).
     ///
-    /// The effect pauses decoration-chain animation while idle. That is the only
-    /// lever that lets the GPU leave its top performance state: an animated pack
-    /// repaints every window carrying it on every vsync, and it is the EXISTENCE of
-    /// per-frame work, not its size, that holds the clocks up.
+    /// The effect pauses decoration-chain animation while idle, the only lever
+    /// that lets the GPU leave its top performance state. An animated pack
+    /// repaints every window carrying it on every vsync, and it is the EXISTENCE
+    /// of per-frame work, not its size, that holds the clocks up.
     std::unique_ptr<PhosphorServiceIdle::IdleService> m_idleService;
 
-    /// Coalesces idle-ladder rebuilds. Rearming is not free and not silent — it
-    /// destroys and recreates the compositor's ext-idle-notify-v1 object, and while
-    /// the session is idle it announces a resume, which wakes every decorated window.
-    /// The "Idle after" slider writes on every step of a drag, so the rebuild is
-    /// deferred to one net reconfigure once the value settles.
+    /// Coalesces idle-ladder rebuilds. Rearming destroys and recreates the
+    /// compositor's ext-idle-notify-v1 object, and while idle it announces a
+    /// resume that wakes every decorated window. The "Idle after" slider writes
+    /// on every drag step, so rebuilds defer to one net reconfigure.
     QTimer m_idleStagesRefreshTimer;
     static constexpr int kIdleStagesRefreshDebounceMs = 250;
 
-    /// Retry budget for an idle ladder that would not arm. The failure this covers is a
-    /// login race (the seat's input devices are not advertised yet), so it resolves in well
-    /// under a second or it is not going to resolve at all — a handful of tries a second
-    /// apart is generous. When the budget runs out the feature degrades to off, which is
-    /// what it silently did before anyone was checking.
+    /// Retry budget for an idle ladder that would not arm. This covers a login
+    /// race (the seat's input devices are not advertised yet), which resolves in
+    /// well under a second or not at all, so a few tries a second apart is
+    /// generous. Exhausting it degrades the feature to off, which is what it
+    /// silently did before anyone was checking.
     static constexpr int kIdleArmRetries = 5;
     static constexpr int kIdleArmRetryDelayMs = 1000;
     int m_idleArmRetriesLeft = kIdleArmRetries;
@@ -853,6 +837,12 @@ private:
     /// a triple-wide disconnect would silently delete those too. Qt::UniqueConnection
     /// is not an option here: it does not apply to lambda/functor connections.
     QList<QMetaObject::Connection> m_restartScopedConnections;
+    /// Handles for the autotile shortcut connections installed by
+    /// initializeAutotile(). Separate from m_restartScopedConnections because
+    /// that list is cleared in connectLayoutSignals(), which start() calls
+    /// AFTER initializeAutotile() — sharing one list would drop these handles
+    /// the moment after they were installed.
+    QList<QMetaObject::Connection> m_autotileShortcutConnections;
 
     std::unique_ptr<PhosphorWorkspaces::VirtualDesktopManager> m_virtualDesktopManager;
     std::unique_ptr<PhosphorWorkspaces::ActivityManager> m_activityManager;
